requirements.txt
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
#python>=3.10
matplotlib>=3.2.2
numpy<1.27,>=1.22 
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=2.2.0,<=2.5.0
torchvision>=0.13.0
tqdm>=4.41.0
protobuf<4.21.3
urllib3<2.1,>=1.25.4
packaging>=21.0

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas<2,>=1.1.2 
seaborn>=0.11.0
networkx~=2.5 

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow
openpyxl
xlsxwriter
